Reduce diffs.

+++ cddl/contrib/opensolaris/tools/ctf/cvt/alist.c
@@ -55,7 +55,7 @@ static int
 alist_hash(int nbuckets, void *arg)
 {
 	alist_el_t *el = arg;
-	int num = (int)(uintptr_t)el->ale_name;
+	size_t num = (size_t)el->ale_name;
 
 	return (num % nbuckets);
 }
@@ -65,7 +65,7 @@ alist_cmp(void *arg1, void *arg2)
 {
 	alist_el_t *el1 = arg1;
 	alist_el_t *el2 = arg2;
-	return ((uintptr_t)el1->ale_name != (uintptr_t)el2->ale_name);
+	return (el1->ale_name != el2->ale_name);
 }
 
 alist_t *
